requirements.txt
Flask
flask-cors
flask-jwt-extended
flask-sqlalchemy
psycopg2
bcrypt
python-dotenv
twilio
sendgrid
